requirements.txt
blinker==1.9.0
click==8.2.1
cloud-sql-python-connector==1.7.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-cloud-secret-manager==2.20.0
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pg8000==1.31.2
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
